LayoutTests/platform/glib/imported/w3c/web-platform-tests/mathml/relations/html5-tree/dynamic-childlist-002-expected.txt


FAIL Appending and removing children to maction assert_approx_equals: block position (child 0) expected -26 +/- 1 but got -70
PASS Appending and removing children to menclose
PASS Appending and removing children to merror
FAIL Appending and removing children to mfrac assert_approx_equals: block position (child 0) expected 0 +/- 1 but got 3.234375
FAIL Appending and removing children to mi assert_approx_equals: inline position (child 0) expected -22.34375 +/- 1 but got -8
FAIL Appending and removing children to mmultiscripts assert_approx_equals: inline size expected 0 +/- 1 but got 43.5625
FAIL Appending and removing children to mn assert_approx_equals: block position (child 0) expected -1082 +/- 1 but got -1100
FAIL Appending and removing children to mo assert_approx_equals: inline position (child 0) expected -87.875 +/- 1 but got -8
FAIL Appending and removing children to mover assert_approx_equals: inline size expected 10 +/- 1 but got 0
PASS Appending and removing children to mpadded
PASS Appending and removing children to mphantom
FAIL Appending and removing children to mroot assert_approx_equals: inline position (child 0) expected 0 +/- 1 but got 4.4375
PASS Appending and removing children to mrow
FAIL Appending and removing children to ms assert_approx_equals: block position (child 0) expected -39053 +/- 1 but got -39071
PASS Appending and removing children to mspace
FAIL Appending and removing children to msqrt assert_approx_equals: inline size expected 13 +/- 1 but got 16
PASS Appending and removing children to mstyle
FAIL Appending and removing children to msub assert_approx_equals: block position (child 0) expected 0 +/- 1 but got -5.5
FAIL Appending and removing children to msubsup assert_approx_equals: block position (child 0) expected 0 +/- 1 but got 2.5
FAIL Appending and removing children to msup assert_approx_equals: inline size expected 10.890625 +/- 1 but got 0
PASS Appending and removing children to mtable
FAIL Appending and removing children to mtext assert_approx_equals: block position (child 0) expected -10046162 +/- 1 but got -10046180
FAIL Appending and removing children to munder assert_approx_equals: block size expected 22 +/- 1 but got 24
FAIL Appending and removing children to munderover assert_approx_equals: inline size expected 0 +/- 1 but got 10
PASS Appending and removing children to semantics
maction:
mfrac:
mi:
mmultiscripts:
mn:
mo:
mover:
mroot:
ms:
msqrt:
msub:
msubsup:
msup:
mtext:
munder:
munderover:
